Renamed plugin package and plugin name to influxdb
