+++ influxdb/influxdb.go
@@ -17,7 +17,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package influx
+package influxdb
 
 import (
 	"bytes"
