+++ influxdb/influxdb_small_test.go
@@ -18,7 +18,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package influx
+package influxdb
 
 import (
 	"testing"
